chore: move terminal related code to application package
