+++ server/application/terminal_test.go
@@ -1,4 +1,4 @@
-package terminal
+package application
 
 import (
 	"testing"
